requirements.txt
gradio
torch
torchvision
pillow
numpy